backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.3
pydantic[email]>=2.7.0
pydantic-settings>=2.4.0
PyJWT>=2.8.0
httpx==0.27.2
google-auth==2.35.0
cachecontrol==0.14.0
slowapi>=0.1.9
pytest==8.3.0
pytest-asyncio==0.24.0
